backend/requirements.txt
amqp==5.3.1
anyio==4.7.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.3.6
cffi==1.17.1
channels==4.0.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==42.0.8
cssselect2==0.7.0
Django==5.0.10
django-celery-beat==2.7.0
django-extensions==3.2.3
django-filter==23.3
django-guardian==2.4.0
django-ledger==0.5.6.2
django-prometheus==2.3.1
django-timezone-field==7.0
django-treebeard==4.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.7.0
drf-nested-routers==0.94.1
drf-spectacular==0.26.5
Faker==33.1.0
Flask==3.1.0
fonttools==4.55.3
future==1.0.0
gunicorn==21.2.0
h11==0.14.0
hl7==0.4.5
html5lib==1.1
httptools==0.6.4
idna==3.10
impacket==0.12.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.4.2
ldap3==2.9.1
ldapdomaindump==0.9.4
Markdown==3.7
MarkupSafe==3.0.2
msgpack==1.1.0
ofxtools==0.9.5
packaging==24.2
Pillow==10.1.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.21.0
pydyf==0.11.0
PyJWT==2.8.0
pyOpenSSL==24.0.0
pyphen==0.17.0
pyserial==3.5
pysmb==1.2.9.1
pytest==8.3.4
pytest-django==4.9.0
pytest-mock==3.14.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
reportlab==4.0.8
rpds-py==0.10.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.4.4
text-unidecode==1.3
tinycss2==1.4.0
tqdm==4.67.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
uvicorn==0.27.1
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.3
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
websockets==14.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
google-generativeai==0.3.2
django-cors-headers==4.3.1
openpyxl==3.1.2